pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["haze518 <ashrbah@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = {extras = ["debug-server"], version = "^0.70.0"}
alembic = "^1.6.5"
databases = "^0.4.3"
psycopg2-binary = "^2.9.1"
asyncpg = "^0.23.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
SQLAlchemy-Utils = "^0.37.8"
pytest-asyncio = "^0.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
